+++ inc/V4l2Capturer.h
@@ -18,8 +18,9 @@
 #include "EncodedVideoFrameBuffer.h"
 
 #include "V4l2Capture.h"
+#include "VideoSource.h"
 
-class V4l2Capturer : public rtc::VideoSourceInterface<webrtc::VideoFrame>
+class V4l2Capturer : public VideoSource
 {
 public:
 	static V4l2Capturer *Create(const std::string &videourl, const std::map<std::string, std::string> &opts, std::unique_ptr<webrtc::VideoDecoderFactory> &videoDecoderFactory)
@@ -68,7 +69,7 @@ private:
 	{
 		m_width = width;
 		m_height = height;
-		
+
 		std::string device = "/dev/video0";
 		if (videourl.find("v4l2://") == 0) {
 			device = videourl.substr(strlen("v4l2://"));
@@ -156,20 +157,9 @@ private:
 		}
 	}
 
-	void AddOrUpdateSink(rtc::VideoSinkInterface<webrtc::VideoFrame> *sink, const rtc::VideoSinkWants &wants) override
-	{
-		m_broadcaster.AddOrUpdateSink(sink, wants);
-	}
-
-	void RemoveSink(rtc::VideoSinkInterface<webrtc::VideoFrame> *sink) override
-	{
-		m_broadcaster.RemoveSink(sink);
-	}
-
 	bool                                                    m_stop;
 	std::thread                                             m_capturethread;
 	std::unique_ptr<V4l2Capture>                            m_capture;
-	rtc::VideoBroadcaster                                   m_broadcaster;
 	rtc::scoped_refptr<webrtc::EncodedImageBufferInterface> m_sps;
 	rtc::scoped_refptr<webrtc::EncodedImageBufferInterface> m_pps;
 	int                                                     m_width;
