+++ inc/VideoDecoder.h
@@ -271,3 +271,26 @@ class VideoDecoder : public rtc::VideoSourceInterface<webrtc::VideoFrame>, publi
         int64_t                               m_prevts;
 
 };
+
+class VideoSourceWithDecoder : public rtc::VideoSourceInterface<webrtc::VideoFrame>
+{
+public:
+	VideoSourceWithDecoder(const std::map<std::string,std::string> & opts, std::unique_ptr<webrtc::VideoDecoderFactory>& videoDecoderFactory, bool wait = false): m_decoder(opts, videoDecoderFactory, wait) {}
+
+    // overide rtc::VideoSourceInterface<webrtc::VideoFrame>
+    void AddOrUpdateSink(rtc::VideoSinkInterface<webrtc::VideoFrame> *sink, const rtc::VideoSinkWants &wants)
+    {
+        m_decoder.AddOrUpdateSink(sink, wants);
+    }
+
+    void RemoveSink(rtc::VideoSinkInterface<webrtc::VideoFrame> *sink)
+    {
+        m_decoder.RemoveSink(sink);
+    }    
+
+	int width() { return m_decoder.width();  }
+    int height() { return m_decoder.height();  } 
+
+protected:
+    VideoDecoder                       m_decoder;	
+};
